requirements.txt
click==6.6
cattle==0.5.4
pyyaml==3.12
